common::impl_pallet_balances! macro
