+++ common/src/lib.rs
@@ -39,6 +39,9 @@ pub mod gas_provider;
 #[cfg(feature = "runtime-benchmarks")]
 pub mod benchmarking;
 
+#[cfg(feature = "std")]
+mod pallet_tests;
+
 use core::fmt;
 use frame_support::{
     codec::{self, Decode, Encode},
