+++ pallets/gear-bank/Cargo.toml
@@ -23,6 +23,7 @@ pallet-authorship.workspace = true
 sp-runtime.workspace = true
 
 [dev-dependencies]
+common = { workspace = true, features = ["std"] }
 primitive-types.workspace = true
 
 # Substrate deps
