+++ pallets/gear-debug/src/mock.rs
@@ -42,17 +42,7 @@ type Balance = u128;
 
 pub const BLOCK_AUTHOR: AccountId = 255;
 
-impl pallet_balances::Config for Test {
-    type MaxLocks = ();
-    type MaxReserves = ();
-    type ReserveIdentifier = [u8; 8];
-    type Balance = Balance;
-    type DustRemoval = ();
-    type RuntimeEvent = RuntimeEvent;
-    type ExistentialDeposit = ExistentialDeposit;
-    type AccountStore = System;
-    type WeightInfo = ();
-}
+common::impl_pallet_balances!(Test);
 
 parameter_types! {
     pub const BlockHashCount: u64 = 250;
