+++ pallets/gear-voucher/Cargo.toml
@@ -35,6 +35,7 @@ pallet-balances.workspace = true
 [dev-dependencies]
 serde.workspace = true
 env_logger.workspace = true
+common = { workspace = true, features = ["std"] }
 sp-core = {workspace = true, features = ["std"] }
 
 [features]
