+++ pallets/payment/Cargo.toml
@@ -34,6 +34,7 @@ pallet-transaction-payment.workspace = true
 [dev-dependencies]
 serde.workspace = true
 env_logger.workspace = true
+common = { workspace = true, features = ["std"] }
 wabt.workspace = true
 gear-core.workspace = true
 sp-io = { workspace = true, features = ["std"] }
